requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
pymongo==4.6.1
python-dotenv==1.0.0
werkzeug==3.0.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
reportlab==4.0.7
bcrypt==4.1.2
email-validator==2.1.0
gunicorn==21.2.0
